Split Event into Event + EventSubscriber
